ref : 구조 변경 및 디자인 변경
