+++ app/login/page.tsx
@@ -1,7 +1,9 @@
+import LoginFormView from "./_view/LoginFormView";
+
 export default function Home() {
   return (
     <main className="flex flex-row h-screen">
-      <div>asd</div>
+      <LoginFormView />
     </main>
   );
 }
